refactor: remove GlassFish prefix from managed and remote classes
Managed module:
- GlassFishManagedContainerConfiguration -> ManagedContainerConfiguration
- GlassFishManagedContainerExtension -> ManagedContainerExtension
- GlassFishManagedDeployableContainer -> ManagedDeployableContainer
- GlassFishServerControl -> ServerControl

Remote module:
- GlassFishExtension -> Extension
- GlassFishRestDeployableContainer -> RestDeployableContainer

Update SPI service files for both modules.
