+++ managed/src/main/java/org/jboss/arquillian/container/glassfish/managed/ManagedContainerExtension.java
@@ -24,9 +24,9 @@ import org.jboss.arquillian.core.spi.LoadableExtension;
  * @author <a href="mailto:aslak@redhat.com">Aslak Knutsen</a>
  * @version $Revision: $
  */
-public class GlassFishManagedContainerExtension implements LoadableExtension {
+public class ManagedContainerExtension implements LoadableExtension {
     @Override
     public void register(ExtensionBuilder builder) {
-        builder.service(DeployableContainer.class, GlassFishManagedDeployableContainer.class);
+        builder.service(DeployableContainer.class, ManagedDeployableContainer.class);
     }
 }
