+++ managed/src/main/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,1 @@
-org.jboss.arquillian.container.glassfish.managed.GlassFishManagedContainerExtension
+org.jboss.arquillian.container.glassfish.managed.ManagedContainerExtension
