+++ remote/src/main/java/org/jboss/arquillian/container/glassfish/remote/Extension.java
@@ -26,9 +26,9 @@ import org.jboss.arquillian.core.spi.LoadableExtension;
  * @author <a href="mailto:aslak@redhat.com">Aslak Knutsen</a>
  * @version $Revision: $
  */
-public class GlassFishExtension implements LoadableExtension {
+public class Extension implements LoadableExtension {
     @Override
     public void register(ExtensionBuilder builder) {
-        builder.service(DeployableContainer.class, GlassFishRestDeployableContainer.class);
+        builder.service(DeployableContainer.class, RestDeployableContainer.class);
     }
 }
